Changed restrict to __restrict and used a common EGL configuration
In order to avoid some issues with some C++ compilers, restrict is
renamed __restrict. However, C++ compilers can't understand C struct
and array initialisers so... I might revert this in the future.

Used a common EGL configuration system, using macros. This should
make it easier to correct and customise.

Preparing to merge with the recent changes.

Signed-off-by: Myy <myy@miouyouyou.fr>