Use XCB instead of Xlib and added quads structures helpers funcs
Push a key, release it and watch for Xlib registering only the key
press ! Making input management, a fucking disaster !

So Xlib was dropped in favor of XCB, which tends to do the trick for
now, except that the destroy handling is missing. The window closes
fine but the program is not alerted by this action, making the whole
point moot...
So the X11 implementation is semi-working, semi-broken. Like my screen
actually.

Meanwhile, new functions for Unsigned Short quads were added.
* US_two_tris_quad_3D_store generates the quad coordinates using the
  informations provided in arguments, and store them in the provided
  US_two_tris_quad_3D * pointer.

* US_two_tris_quad_3D_draw draw an array of such quads, using OpenGL
  functions. The array must have been stored in an OpenGL buffer
  before, and the buffer ID must be passed to the function.

Signed-off-by: Myy <myy@miouyouyou.fr>