Quick bugfix for the new X11/XCB platform support
[Bugfix] The program now terminates properly when the window is
close.
I really appreciate the nice documentation provided by XCB, which
consists in searching endlessly on the web for a few lines of code
disseminated between freedesktop mailing lists, and Vulkan support
documentation...

* myy_single_string_to_quads converts a single string to quads
  and returns the size of the converted quads.

The strings to quads functions will soon be rewritten to return
the number of quads generated AND the size, in order to store them
AND draw them efficiently.

Signed-off-by: Myy <myy@miouyouyou.fr>