First commit for the C++ branch of these helpers
Due to being tired of retyping the name of the object type AND name,
and generating names with very long names to avoid naming conflicts,
I decided to use C++ for the Graphical Assembly project.

C++ has its issues, notably when it comes to name mangling and the
whole 'whiny' attitude of the compiler.
Due to these problems, it is extremely difficult to make a project
with and C++ files coexisting. Therefore, I had to port these helpers
to C++.

Currently, only the X11/XCB part was tested.

Signed-off-by: Myy <myy@miouyouyou.fr>