Switching to a weak symbol system.
Weak symbols are basically symbols that can be overridden by standard
symbols (which are the "strong" symbols here).

That allows the user myy.c (or whatever main C file he defines) to
only contain what he wants to use, leaving the rest to default
implementations.

I find this so convenient that I don't think I'll go back to the old
way, unless Android has issues with that... I'll have to check that.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>