Moved the platform handling to the stubs...
I could replace this functions pointers with standard function calls.
However, most of the "stop" functions stop the main window draw loop
by setting a "running" variable to 0.

If done through function calls, the "running" variable would have to
be global (since you'll need to access it through another function,
that receives no argument).
Which means that the loop will have to constantly access some heap
data to check whether it's at 0 or 1.
The drawing function being one of the hottest point of execution,
being called "at least" 60 times per second, it needs to avoid
random heap accesses.

With the function pointers, the "running" variable can be on the
stack.
Then the "pointer" to this variable will be passed to the stop
function.
I expect the stack to be constantly cached in memory for various
reasons, so reading it again and again should always lead to
cache hits.

Of course, without proper testing, these are just wild hypothesis.

Also, there could be another way to handle this, like sending a
signal to the window draw thread, then catch the signal in the
thread and "quit" peacefully.
However that sounds like way too much work for barely any gain.

So, yeah, I'll keep the function pointers at the moment. It makes
design a little cleaner on the critical parts.

Meanwhile, I added a TODO list.

By rereading the code, it seems that we're still dealing with
random handcoded resolutions and window names since no way to
provide these information have been added.

The TODO list should serve as a global remainder of what to do,
obviously, since my current IDE do not parse the TODO of each C file,
and I'm lazy so "grepping" for TODO with "-A4" parameters to
scan for remaining TODO is a terrible way IMHO.

At last, I removed some empty lines with only spaces in it.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>