Is (u)int_fastN_t THAT FAST ?
I mean... there might be some cases where it is... if you REALLY
want a type with that much bits "at best".
Note that if you just want some "register size", to store some
random values, and want to say to the developers "it's not going
to go over 256", you'll still be better with uint32_t or uint64_t.
The reason being that uint_fastN_t will generate "and (N_MASK)"
instructions everywhere, in order to ensure that you stay in the
limits.

Compared to a simple uint32_t or uint64_t, fitted to the
architecture, that will not generate additional "and" instructions,
I'll take the uint32_t or uint64_t.
For that purpose, I made (u)intreg_t and the C bool uses this
type now.

The C bool is not suited for storage. If you want to define
data structure, better use uint8_t or whatever is better suited
for the operation.

* myy_gl_text_infos_chars_to_quads has been added.
  It behaves almost identically to myy_strings_list_to_quads,
  except that the string size is defined explicitly, instead of
  using strlen to "guess" it.
  Note that all the gl_text might be renamed to gl_text_atlas,
  since it's what it's about.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>