Repaired the Android implementation
Which reminds me that compiling with another compiler allows you to
spot several bugs here and there, that didn't appear with another.
Well, I don't enable "-Wall" with GCC nor Clang, so there might be
a lot more issues.
At the same time, I don't want to fight with compilers when the
"warning" is really dubious, due to how C syntax and hints limitations.

Anyway, a new function was added for loading assets, because
crawling within the filesystem in order to find some random
asset is NOT the way to go with Android. And should be avoided
anyway.

* myy_asset_map_to_memory(relative_filepath, asset_type);
  Allows you to load an asset, using its relative filepath
  from its supposed folder, and the type of assets, which
  will determine where the actual filepath of the asset.
  Currently only the type : myy_asset_type_texture is
  supported.
  Example :
  myy_asset_map_to_memory("tex.myyraw", myy_asset_type_texture)
  will load map "textures/tex.myyraw" into memory.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>