Random logs and temporary measures to deal with Native Insanity
AMotionEvent_getEventTime led to some crashes at some point,
when the Soft Input keyboard was triggered in the application,
and I continued to handle touch input in my application.

Some logs were added in order to know which Thread invokes
myyTextInputStopped compared to which one calls myy_trigger_text_input.
I'll have to find a simple thread message passing mechanism,
in order to tell the application that the input stopped and
where the input string is located...

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>