The android input works ! Thanks to a lot of horrendous hacks...
That I need to clean up now.

However, yeah, turns out that the multi-threaded systems brought
up by Android -- because they LOVE to throw you on the UI thread
every time -- can bite you in the ass.

When the Java activity call up a native function, it will do it
from its thread, which is obviously different from the native
'UI thread' created by android_app_glue.c.
HOWEVER, since the 'OpenGL context' is created on the 'native
UI thread', you need to abuse the current message passing
mechanism that CANNOT PASS A SINGLE ARGUMENT to the commands...

I guess that's it's for efficiency that they avoid passing
arguments to the commands, since none seem to need them...
So, the best plan might to add a new Looper and pass "myy"
commands from this looper... That seems like nice idea...
And would avoid a ton of dangling references stored here
and there.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>